Demonstration of how to run PgAdmin 4 on GOV.UK PaaS
`cf conduit` is a useful tool for connecting to Postgres backing services, but
it isn't meant for long-lived connectivity. A better solution for that is to
run PgAdmin as an app. This has the dual benefits of being accessible outside
of a developers machine, and being suitable for long-lived connections.

The demonstration is suitable as a basis for tenants to customise to suit their
needs.