Update readme to reflect the name change
We're named "paas-incubator" to "paas-community". We hope this repo will grow to contain/link to useful work that our users have done, as well as the GOV.UK PaaS team ourselves. This commit updates the readme to state the new purpose. We haven't added very many generic examples to this repo yet, so I'll start encouraging PaaS's engineers to see if they have useful examples to go in here.

Getting community engagement will be a gradual process. We can start linking to this from the documentation once it's a little more structured and we have some useful tenant examples linked.